Add typings for EditorOptions/EditorEvents, remove index signatures
The index signatures which were attached to IAceEditorProps and the
ReactAce class were causing consumers to lose type safety when using
the component. By excplicitly defining the allowed editor options which
are passed down to AceEditor, typescript can be sure we're attempting
to access valid properties on our class rather than defining a catch-all
index signature.