Fix prettier configuration.
The current prettier configuration is not consistent with the formatting
used throughout the project. As a result, running the formatter on a
fresh repository results in a large number of spurious formatting
changes. This commit adjust the formatting rules to the point where
running the formatter only results in a single small change to
src/types.ts, which is also included in the commit.

In addition, because the formatting options are passed on the command
line through npm run, this prevents tools like VSCode from understanding
the project formatting guidelines. Moving the prettier configuration
into package.json enables integration with third-party tools.