Enable simple SSR.
react-ace is almost SSR ready. The main render function merely inserts
an empty div, which is then populated by ace in componentDidMount.
Because componentDidMount is not fired during SSR, SSR should result in
an empty div. Ideally we would start with the content that would be
present after the initial call to editor.resize(), but an empty div is
better than failing.

Currently react-ace fails during SSR because require("ace-builds")
fails. ace-builds expects window to exist, but only to attach a global
ace object to. react-ace doesn't actually use that global ace instance,
rather it uses the handle to ace returned by require. And, again,
because all of the work is done in componentDidMount, just not failing
during SSR doesn't require mocking up much of any kind of realistic
window object like jsdom might try to do. It turns out that a simple
empty object suffices.

Essentially here's what currently happens:

// Blows up during SSR since window is not defined
if ((window as any).ace) {

So let's add an SSR check:

if (typeof window === "undefined") {
  // Now this blows up because it wants to set window.ace to the ace
  // instance
  ace = require("ace-builds");

So instead we can simply do this:

if (typeof window === "undefined") {
  global.window = {}
  ace = require("ace-builds");
  delete global.window;

And all is right with the world.

While this may seem like a trivial change, currently react-ace cannot be
used out of the box with tools like Gatsby that do SSR for production
builds. Workarounds exist using libraries like "@loadable/component",
but why bother with another dependency when the solution is fairly
simple?